updated project settings for xcode 7

+++ GMGridView/GMGridView.m
@@ -787,9 +787,9 @@ static const UIViewAnimationOptions kDefaultAnimationOptions = UIViewAnimationOp
 
 - (void)sortingMoveDidContinueToPoint:(CGPoint)point
 {
-    int position = [self.layoutStrategy itemPositionFromLocation:point];
+    NSInteger position = [self.layoutStrategy itemPositionFromLocation:point];
     
-    int tag = MIN(position,_numberTotalItems-1) + kTagOffset;
+    NSInteger tag = MIN(position,_numberTotalItems-1) + kTagOffset;
     
     if (position != GMGV_INVALID_POSITION && position != _sortFuturePosition) 
     {
@@ -1655,7 +1655,7 @@ static const UIViewAnimationOptions kDefaultAnimationOptions = UIViewAnimationOp
     {        
         cell = [self newItemSubViewForPosition:index];
         
-        for (int i = _numberTotalItems - 1; i >= index; i--)
+        for (NSInteger i = _numberTotalItems - 1; i >= index; i--)
         {
             UIView *oldView = [self cellForItemAtIndex:i];
             oldView.tag = oldView.tag + 1;
@@ -1708,7 +1708,7 @@ static const UIViewAnimationOptions kDefaultAnimationOptions = UIViewAnimationOp
     
     GMGridViewCell *cell = [self cellForItemAtIndex:index];
     
-    for (int i = index + 1; i < _numberTotalItems; i++)
+    for (NSInteger i = index + 1; i < _numberTotalItems; i++)
     {
         GMGridViewCell *oldView = [self cellForItemAtIndex:i];
         oldView.tag = oldView.tag - 1;
